Rename app from 'wallet' to 'Plasma Wallet'
