+++ android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.wallet",
+    package = "co.omisego.plasmawallet",
 )
 
 android_resource(
     name = "res",
-    package = "com.wallet",
+    package = "co.omisego.plasmawallet",
     res = "src/main/res",
 )
