+++ android/app/bin/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.wallet">
+  package="co.omisego.plasmawallet">
 
     <uses-permission android:name="android.permission.INTERNET" />
