+++ android/app/proguard-rules.pro
@@ -9,7 +9,7 @@
 
 # Add any project specific keep options here:
 
--keep class com.wallet.BuildConfig { *; }
+-keep class co.omisego.plasmawallet.BuildConfig { *; }
 # React Native
 
 # Keep our interfaces so they can be used by other ProGuard rules.
