+++ android/app/src/main/java/co/omisego/plasmawallet/MainActivity.java
@@ -1,4 +1,4 @@
-package com.wallet;
+package co.omisego.plasmawallet;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
@@ -23,6 +23,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "wallet";
+        return "PlasmaWallet";
     }
 }
