+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'wallet'
+rootProject.name = 'PlasmaWallet'
 include ':react-native-randombytes'
 project(':react-native-randombytes').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-randombytes/android')
 include ':react-native-push-notification'
