+++ ios/PlasmaWallet-tvOS/Info.plist
