+++ ios/PlasmaWallet.xcodeproj/wallet.xcodeproj/xcshareddata/xcschemes/wallet.xcscheme
