+++ ios/PlasmaWallet/AppDelegate.h
