+++ ios/PlasmaWallet/Images.xcassets/AppIcon.appiconset/Contents.json
