+++ ios/PlasmaWallet/Images.xcassets/Contents.json
