+++ ios/PlasmaWalletTests/Info.plist
