+++ ios/PlasmaWalletTests/walletTests/walletTests.m
