+++ ios/Podfile
@@ -1,8 +1,8 @@
 platform :ios, '10.0'
 require_relative '../node_modules/@react-native-community/cli-platform-ios/native_modules'
 
-target 'wallet' do
-  # Pods for wallet
+target 'PlasmaWallet' do
+  # Pods for PlasmaWallet
   pod 'React', :path => '../node_modules/react-native/'
   pod 'React-Core', :path => '../node_modules/react-native/React'
   pod 'React-DevSupport', :path => '../node_modules/react-native/React'
@@ -38,7 +38,7 @@ target 'wallet' do
 
   pod 'react-native-randombytes', :path => '../node_modules/react-native-randombytes'
 
-  target 'walletTests' do
+  target 'PlasmaWalletTests' do
     inherit! :search_paths
     # Pods for testing
   end
@@ -61,10 +61,10 @@ post_install do |installer|
   end
  end
 
-target 'wallet-tvOS' do
-  # Pods for wallet-tvOS
+target 'PlasmaWallet-tvOS' do
+  # Pods for PlasmaWallet-tvOS
 
-  target 'wallet-tvOSTests' do
+  target 'PlasmaWallet-tvOSTests' do
     inherit! :search_paths
     # Pods for testing
   end
