+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "wallet",
+  "name": "PlasmaWallet",
   "version": "0.0.1",
   "private": true,
   "scripts": {
